requirements.txt
django==4.1.4
python-dotenv==1.0.0
Pillow==9.3.0
psycopg2==2.9.6
aiogram==2.25.1
requests==2.29.0
isort==5.12.0
djangorestframework==3.14.0
